requirements.txt
numpy>=1.2
opencv-python>=4.4
Pillow>=8.1
scikit-image>=0.18
